requirements.txt
Pillow
matplotlib
pandas
seaborn
streamlit